refactor(api): move challenge-serializer.js to shared scope
