+++ api/src/school/application/assessment-controller.js
@@ -1,4 +1,4 @@
-import * as challengeSerializer from '../../../lib/infrastructure/serializers/jsonapi/challenge-serializer.js';
+import * as challengeSerializer from '../../shared/infrastructure/serializers/jsonapi/challenge-serializer.js';
 import { usecases } from '../shared/usecases/index.js';
 import * as assessmentSerializer from '../../../src/school/infrastructure/serializers/assessment.js';
