+++ api/src/shared/infrastructure/serializers/jsonapi/challenge-serializer.js
